Comprueba si existen los esquemas de BBDD antes de borrarlos

+++ ritsback/resources/db/ddl/00.drop_tables.sql
@@ -1,8 +1,39 @@
 SET client_min_messages TO WARNING;
-DROP SCHEMA conf CASCADE;
-DROP SCHEMA static CASCADE;
-DROP SCHEMA master CASCADE;
 
-CREATE SCHEMA master;
+DO $$
+BEGIN
+    IF EXISTS (
+        SELECT schema_name 
+        FROM information_schema.schemata 
+        WHERE schema_name = 'conf'
+    ) THEN
+        DROP SCHEMA conf CASCADE;
+    END IF;
+END $$;
+
+DO $$
+BEGIN
+    IF EXISTS (
+        SELECT schema_name 
+        FROM information_schema.schemata 
+        WHERE schema_name = 'static'
+    ) THEN
+        DROP SCHEMA static CASCADE;
+    END IF;
+END $$;
+
+
+DO $$
+BEGIN
+    IF EXISTS (
+        SELECT schema_name 
+        FROM information_schema.schemata 
+        WHERE schema_name = 'master'
+    ) THEN
+        DROP SCHEMA master CASCADE;
+    END IF;
+END $$;
+
 CREATE SCHEMA conf;
 CREATE SCHEMA static;
+CREATE SCHEMA master;
