+++ ritsfront/resources/db/ddl/00.drop_tables.sql
@@ -1,4 +1,13 @@
 SET client_min_messages TO WARNING;
-DROP SCHEMA ui CASCADE;
+DO $$
+BEGIN
+    IF EXISTS (
+        SELECT schema_name 
+        FROM information_schema.schemata 
+        WHERE schema_name = 'ui'
+    ) THEN
+        DROP SCHEMA ui CASCADE;
+    END IF;
+END $$;
 
 CREATE SCHEMA ui;
